# docs/swagger.json
{
    "swagger": "2.0",
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "info": {
        "description": "## A simple, yet customizable and incrementally-adoptable backend server for a fully-functional social network",
        "title": "Yappy API Specifications",
        "version": "1.0"
    },
    "paths": {
...
        "/chats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint retrieves a paginated list of the current user chats`",
                "tags": [
                    "Chat"
                ],
                "summary": "Retrieve User Chats",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ChatsResponseSchema"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint sends a message`\n\n`You must either send a text or a file or both.`\n\n`If there's no chat_id, then its a new chat and you must set username and leave chat_id`\n\n`If chat_id is available, then ignore username and set the correct chat_id`\n\n`The file_upload_data in the response is what is used for uploading the file to cloudinary from client`",
                "tags": [
                    "Chat"
                ],
                "summary": "Send a message",
                "parameters": [
                    {
                        "description": "Message object",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageCreateSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageCreateResponseSchema"
                        }
                    }
                }
            }
        },
        "/chats/groups/group": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint creates a group chat.`\n\n`The users_entry field should be a list of usernames you want to add to the group.`\n\n`Note: You cannot add more than 99 users in a group (1 owner + 99 other users = 100 users total).`",
                "tags": [
                    "Chat"
                ],
                "summary": "Create a Group Chat",
                "parameters": [
                    {
                        "description": "Chat object",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupChatCreateSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupChatInputResponseSchema"
                        }
                    }
                }
            }
        },
        "/chats/messages/{message_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint updates a message.`\n\n`You must either send a text or a file or both.`\n\n`The file_upload_data in the response is what is used for uploading the file to cloudinary from client.`",
                "tags": [
                    "Chat"
                ],
                "summary": "Update a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID (uuid)",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message object",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageUpdateSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.MessageCreateResponseSchema"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint deletes a message.`",
                "tags": [
                    "Chat"
                ],
                "summary": "Delete a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID (uuid)",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSchema"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint retrieves all messages in a chat`",
                "tags": [
                    "Chat"
                ],
                "summary": "Retrieve messages from a Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID (uuid)",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ChatResponseSchema"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint deletes a group chat.`",
                "tags": [
                    "Chat"
                ],
                "summary": "Delete a Group Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID (uuid)",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseSchema"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint updates a group chat.`",
                "tags": [
                    "Chat"
                ],
                "summary": "Update a Group Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID (uuid)",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chat object",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupChatInputSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GroupChatInputResponseSchema"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "choices.ChatTypeChoice": {
            "type": "string",
            "enum": [
                "DM",
                "GROUP"
            ],
            "x-enum-varnames": [
                "CDM",
                "CGROUP"
            ]
        },
...
        "models.Chat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "ctype": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/choices.ChatTypeChoice"
                        }
                    ],
                    "example": "DM"
                },
                "description": {
                    "type": "string",
                    "example": "A nice group for tech enthusiasts"
                },
                "file_upload_data": {
                    "$ref": "#/definitions/utils.SignatureFormat"
                },
                "id": {
                    "type": "string",
                    "example": "d10dde64-a242-4ed0-bd75-4c759644b3a6"
                },
                "image": {
                    "type": "string",
                    "example": "https://img.url"
                },
                "latest_message": {
                    "$ref": "#/definitions/models.LatestMessageSchema"
                },
                "name": {
                    "type": "string",
                    "example": "My Group"
                },
                "owner": {
                    "$ref": "#/definitions/models.UserDataSchema"
                },
                "updated_at": {
                    "type": "string"
                },
                "users": {
                    "description": "omitempty later to show for groups",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserDataSchema"
                    }
                }
            }
        },
...
        "models.LatestMessageSchema": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                },
                "sender": {
                    "$ref": "#/definitions/models.UserDataSchema"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file": {
                    "type": "string",
                    "example": "https://img.url"
                },
                "file_upload_data": {
                    "$ref": "#/definitions/utils.SignatureFormat"
                },
                "id": {
                    "type": "string",
                    "example": "d10dde64-a242-4ed0-bd75-4c759644b3a6"
                },
                "sender": {
                    "$ref": "#/definitions/models.UserDataSchema"
                },
                "text": {
                    "type": "string",
                    "example": "Jesus is King"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.UserDataSchema": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://img.url"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "john-doe"
                }
            }
        },
...
        "schemas.ChatResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.MessagesSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "schemas.ChatsResponseDataSchema": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Chat"
                    }
                },
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "last_page": {
                    "type": "integer",
                    "example": 100
                },
                "per_page": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "schemas.ChatsResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.ChatsResponseDataSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "schemas.GroupChatCreateSchema": {
            "type": "object",
            "required": [
                "name",
                "usernames_to_add"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is a group for bosses."
                },
                "file_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Dopest Group"
                },
                "usernames_to_add": {
                    "type": "array",
                    "maxItems": 99,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "john-doe"
                    ]
                }
            }
        },
        "schemas.GroupChatInputResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Chat"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "schemas.GroupChatInputSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is a group for bosses."
                },
                "file_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Dopest Group"
                },
                "usernames_to_add": {
                    "type": "array",
                    "maxItems": 99,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "john-doe"
                    ]
                },
                "usernames_to_remove": {
                    "type": "array",
                    "maxItems": 99,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "john-doe"
                    ]
                }
            }
        },
...
        "schemas.MessageCreateResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Message"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "schemas.MessageCreateSchema": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string",
                    "example": "d10dde64-a242-4ed0-bd75-4c759644b3a6"
                },
                "file_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "text": {
                    "type": "string",
                    "example": "I am not in danger skyler, I am the danger"
                },
                "username": {
                    "type": "string",
                    "example": "john-doe"
                }
            }
        },
        "schemas.MessageUpdateSchema": {
            "type": "object",
            "properties": {
                "file_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "text": {
                    "type": "string",
                    "example": "The Earth is the Lord's and the fullness thereof"
                }
            }
        },
        "schemas.MessagesResponseDataSchema": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "last_page": {
                    "type": "integer",
                    "example": 100
                },
                "per_page": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "schemas.MessagesSchema": {
            "type": "object",
            "properties": {
                "chat": {
                    "$ref": "#/definitions/models.Chat"
                },
                "messages": {
                    "$ref": "#/definitions/schemas.MessagesResponseDataSchema"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserDataSchema"
                    }
                }
            }
        },
...
        "schemas.ResponseSchema": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.SignatureFormat": {
            "type": "object",
            "properties": {
                "public_id": {
                    "type": "string",
                    "example": "images/f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "signature": {
                    "type": "string",
                    "example": "e1ba4683fbbf90b75ca22e9f8e545b18c6b24eae"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1678828200
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use bearer token to access secured APIs",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}